更新通用框架使之更通用
