+++ program/apps/index/router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 
 from apps.index.views.health import router as health_router
-from apps.index.views.init import router as init_router
+from apps.index.views.db_init_handler.init import router as init_router
 from watchtower.settings import settings
 
 router = APIRouter()
