+++ program/apps/index/views/db_init_handler/business_init.py
@@ -1,0 +1,2 @@
+def run(*args, **kwargs):
+    pass
