+++ program/apps/index/views/db_init_handler/init_db_items.py
